requirements.txt
langchain
boto3
python-dotenv
youtube-transcript-api